[incubator-kie-issues#1206] Small refactoring/improvement
